fix error due to linux brew

+++ tests/bats/03-install-mdm/02-install.bats
@@ -38,7 +38,7 @@ setup() {
 # additional tools part 1
 @test 'launcher after initial install' {
   # remove a inconsequential tool to run this test locally
-  rm "$(brew --prefix)"/etc/bash_completion.d/docker* || :
+  is_mac && rm "$(brew --prefix)"/etc/bash_completion.d/docker* || :
   run "$lib_dir/launcher"
   assert_success
   assert_output -p "additional tools"
